+++ src/integrators/photonmapper/bre.h
@@ -58,10 +58,10 @@ protected:
 	AABB buildHierarchy(size_t index);
 
 	/// Heap convenience routines
-	inline size_t leftChild(size_t index) const { return 2*index; }
-	inline size_t rightChild(size_t index) const { return 2*index + 1; }
-	inline bool isInnerNode(size_t index) const { return index <= m_lastInnerNode; }
-	inline bool hasRightChild(size_t index) const { return index <= m_lastRChildNode; }
+	inline uint32_t leftChild(uint32_t index) const { return 2*index; }
+	inline uint32_t rightChild(uint32_t index) const { return 2*index + 1; }
+	inline bool isInnerNode(uint32_t index) const { return index <= m_lastInnerNode; }
+	inline bool hasRightChild(uint32_t index) const { return index <= m_lastRChildNode; }
 private:
 	struct BRENode {
 		AABB aabb;
